ids568-milestone3-khuzaimaalam2000/requirements.txt
mlflow==2.9.2
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
apache-airflow==2.10.4
pytest==7.4.3